Create MADGraph_Docker
This is a file that has been slightly changed from the one found at "https://hub.docker.com/r/hfukuda/madgraph/dockerfile". I have added the MADAnalysis, as I wish to tun my simulations with the MADAnalysis turned on. 

The intention is to also update the file so that it uses the latest version of MADGraph.