Create RunDockerGenerate.sh
This file contains the commands to run the simulations inside the Docker container. This script will create an instance of the Docker container, and remove it once the simulation has been completed.

Outputs from this script will be saved in the folder Documents\DockerOutputs

It should be written such that the process of making multiple events is automated and that only this script needs to be executed.